Envio

+++ wwwroot/js/telefonos.js
@@ -5,8 +5,7 @@
         
         limpiarTelefonos();
         $('#btnModalTelefono').show();
-        $('#telefonoPersonaID').val($('#personaGeneralID').val());
-        
+        $('#telefonoPersonaID').val($('#personaGeneralID').val());        
     });
 
     $('#btnModalTelefono').click(function () {
@@ -14,30 +13,26 @@
     });
 });
 
-
-
 $('#formTelefonos').submit(function (e) {
     listadoContactos = document.querySelector('#tableTelefonos');
-    console.log("listado", listadoContactos)
     nuevoContacto = document.createElement('tr');
     
     var numero = $('#numeroTelefono').val();
     var CompaniaID = $('#companiaTelefono').val();
     var Tipo_TelefonoID = $('#tipoTelefono').val();
     var persona = $('#personaGeneralID').val();
-    //var telefonoID = $("#telefonoID").val();
-    /*
-    console.log("numero", numero);
-    console.log("telefono",CompaniaID);
-    console.log("tipo telefono",Tipo_TelefonoID);
-    console.log("id persona",persona);
-    */
+    var notificacion;
+
+    if ($('#notPosTelefono').is(':checked')) {
+        notificacion = true;
+    } else if ($('#notNegTelefono').is(':checked')) {
+        notificacion = false;
+    }
 
     if (numero === '' || CompaniaID === '' || Tipo_TelefonoID === '') {
         console.log('FALTAN DATOS');
     } else {
         if (document.getElementById('btnModalTelefono').innerHTML === "Guardar") {
-            
             $.ajax({
                 type: 'POST',
                 url: "/Peticiones/addTelefono",
@@ -45,32 +40,48 @@ $('#formTelefonos').submit(function (e) {
                     numero: numero,
                     CompaniaID: CompaniaID,
                     Tipo_TelefonoID: Tipo_TelefonoID,
-                    persona: persona
+                    persona: persona,
+                    notificacion: notificacion
                 },
-                //contentType: false,
-                //processData: false,
                 success: function (data) {
-                    alert('Insertado con el id ' + data);
-                    $('#modalTelefonos').modal('hide');
-                    //obtenerDom(data);
-                    console.log("id de la persona", persona);
-                    console.log("id del telefono", data);
+                    if (data === '') {
+                        Swal.fire({
+                            position: 'top-end',
+                            icon: 'error',
+                            title: 'Ya está dado de alta',
+                            showConfirmButton: false,
+                            timer: 1500
+                        })
+                    } else {
+                        Swal.fire({
+                            position: 'top-end',
+                            icon: 'success',
+                            title: 'Guardado exitoso',
+                            showConfirmButton: false,
+                            timer: 1500
+                        })
+                        $('#modalTelefonos').modal('hide');
+                        $('#lblNoTelefonos').hide();
 
-                    nuevoContacto.setAttribute("id", `${data}`);
+                        nuevoContacto.setAttribute("id", `${data}`);
+                        nuevoContacto.setAttribute("class", `tablaTelefono-${data} tablaTelefono`);
 
-                    nuevoContacto.innerHTML = `
-                        <td>${numero}</td>
-                        <td>${CompaniaID}</td>
-                        <td>${Tipo_TelefonoID}</td>
-                        <td>
-                            <button type="button" onclick="editTelefono(this)" data-toggle="modal" data-target="#modalTelefonos" class="btn btn-success" value=${data} name=${persona} id="editTelefono">Editar</button>
-                            <button class="btn btn-primary" onclick="detalleTelefono(this)" data-toggle="modal" data-target="#modalTelefonos" value=${data} name=${persona}>Detalles</button>
-                            <button class="btn btn-danger" onclick="deleteTelefono(this)" value=${data}>Borrar</button>
-                        </td>
-                    `;
+                        var comp = $('#companiaTelefono option[value=' + CompaniaID + ']').text();
+                        var tipo = $('#tipoTelefono option[value=' + Tipo_TelefonoID + ']').text();
 
-                    listadoContactos.appendChild(nuevoContacto);
-                    
+                        nuevoContacto.innerHTML = `
+                            <td>${numero}</td>
+                            <td>${comp}</td>
+                            <td>${tipo}</td>
+                            <td>
+                                <button type="button" onclick="editTelefono(this)" data-toggle="modal" data-target="#modalTelefonos" class="btn btn-success" value=${data} name=${persona} id="editTelefono">Editar</button>
+                                <button type="button" class="btn btn-primary" onclick="detalleTelefono(this)" data-toggle="modal" data-target="#modalTelefonos" value=${data} name=${persona}>Detalles</button>
+                                <button type="button" class="btn btn-danger" onclick="deleteTelefono(this)" value=${data}>Borrar</button>
+                            </td>
+                        `;
+
+                        listadoContactos.appendChild(nuevoContacto);
+                    }
                 },
                 error: function (r) {
                     console.log(r);
@@ -91,31 +102,50 @@ $('#formTelefonos').submit(function (e) {
                         numero: numero,
                         CompaniaID: CompaniaID,
                         Tipo_TelefonoID: Tipo_TelefonoID,
-                        persona: persona
+                        persona: persona,
+                        notificacion: notificacion
                     },
-                    success: function (data) {
-                        console.log(data);
-                        $('#modalTelefonos').modal('hide');
+                    success: function (data) {                        
+                        if (data === '') {
+                            Swal.fire({
+                                position: 'top-end',
+                                icon: 'error',
+                                title: 'Ya está dado de alta',
+                                showConfirmButton: false,
+                                timer: 1500
+                            })
+                        } else {
+                            Swal.fire({
+                                position: 'top-end',
+                                icon: 'success',
+                                title: 'Actualizado exitoso',
+                                showConfirmButton: false,
+                                timer: 1500
+                            })
+                            $('#modalTelefonos').modal('hide');
 
-                        //listadoContactos.appendChild(nuevoContacto);
-                        var elemento = document.getElementById(id);
-                        $(elemento).remove();
+                            var elemento = document.getElementsByClassName(`tablaTelefono-${id}`);
+                            $(elemento).remove();
 
-                        nuevoContacto.setAttribute("id", `${id}`);
+                            nuevoContacto.setAttribute("id", `${id}`);
+                            nuevoContacto.setAttribute("class", `tablaTelefono-${id} tablaTelefono`);
 
-                        nuevoContacto.innerHTML = `
-                        <td>${numero}</td>
-                        <td>${CompaniaID}</td>
-                        <td>${Tipo_TelefonoID}</td>
-                        <td>
-                            <button type="button" onclick="editTelefono(this)" data-toggle="modal" data-target="#modalTelefonos" class="btn btn-success" value=${id} name=${persona} id="editTelefono">Editar</button>
-                            <button class="btn btn-primary" onclick="detalleTelefono(this)" data-toggle="modal" data-target="#modalTelefonos" value=${id} name=${persona}>Detalles</button>
-                            <button class="btn btn-danger" onclick="deleteTelefono(this)" value=${id}>Borrar</button>
-                        </td>
-                    `;
+                            var comp = $('#companiaTelefono option[value=' + CompaniaID + ']').text();
+                            var tipo = $('#tipoTelefono option[value=' + Tipo_TelefonoID + ']').text();
 
-                     listadoContactos.appendChild(nuevoContacto);
+                            nuevoContacto.innerHTML = `
+                                <td>${numero}</td>
+                                <td>${comp}</td>
+                                <td>${tipo}</td>
+                                <td>
+                                    <button type="button" onclick="editTelefono(this)" data-toggle="modal" data-target="#modalTelefonos" class="btn btn-success" value=${id} name=${persona} id="editTelefono">Editar</button>
+                                    <button class="btn btn-primary" onclick="detalleTelefono(this)" data-toggle="modal" data-target="#modalTelefonos" value=${id} name=${persona}>Detalles</button>
+                                    <button class="btn btn-danger" onclick="deleteTelefono(this)" value=${id}>Borrar</button>
+                                </td>
+                            `;
 
+                            listadoContactos.appendChild(nuevoContacto);
+                        }
                     },
                     error: function (r) {
                         console.log(r);
@@ -132,28 +162,51 @@ function limpiarTelefonos() {
     $("#modalTelefonos textarea").val("");
     $("#modalTelefonos select").val("");
     $("span.Telefonos").hide();
+    document.getElementById('notPosTelefono').checked = false;
+    document.getElementById('notNegTelefono').checked = true;
 }
 
 function deleteTelefono(boton) {
-    var id = boton.value;
+    Swal.fire({
+        title: '¿Desea eliminar el campo permanentemente?',
+        text: "Su decisión no podrá ser revertida",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        cancelButtonText: 'Cancelar',
+        confirmButtonText: 'Sí, bórralo'
+    }).then((result) => {
+        if (result.value) {
+            var id = boton.value;
+            var persona = $('#personaGeneralID').val();
 
-    $.ajax({
-        type: 'POST',
-        url: "/Peticiones/deleteTelefono",
-        data: {
-            id: id
-        },
-        success: function (data) {
-            console.log(data);
-            var elemento = document.getElementById(id);
-            $(elemento).remove();
-            //$('#tableTelefonos').load(" #tableTelefonos");
-        },
-        error: function (r) {
-            console.log(r);
+            $.ajax({
+                type: 'POST',
+                url: "/Peticiones/deleteTelefono",
+                data: {
+                    id: id,
+                    persona: persona
+                },
+                success: function (data) {
+                    if (data === '0') {
+                        $('#lblNoTelefonos').show();
+                    }
+                        Swal.fire(
+                            'Borrado exitoso',
+                            'El campo ha sido eliminado',
+                            'success'
+                        )
+                        var elemento = document.getElementsByClassName(`tablaTelefono-${id}`);
+                        $(elemento).remove();
+                    },
+                    error: function (r) {
+                        console.log(r);
+                    }
+                });
+            return false;
         }
-    });
-    return false;
+    })    
 }
 
 function editTelefono(boton) {
@@ -192,13 +245,16 @@ function obtenerTelefono(id) {
             console.log(data);
             $('#TelefonoID').val(data[0]['telefonoID']);
             $('#numeroTelefono').val(data[0]['numero']);
-            var CompaniaID = $('#companiaTelefono').val(data[0]['companiaID']);
-            var Tipo_TelefonoID = $('#tipoTelefono').val(data[0]['tipo_TelefonoID']);
+            $('#companiaTelefono').val(data[0]['companiaID']);
+            $('#tipoTelefono').val(data[0]['tipo_TelefonoID']);
             $('#telefonoPersonaID').val(data[0]['personaID']);
-
-            document.getElementById('companiaTelefono').getElementsByTagName('option')[CompaniaID].selected = 'selected';
-            document.getElementById('tipoTelefono').getElementsByTagName('option')[Tipo_TelefonoID].selected = 'selected';
-
+            if (data[0]['notificacion'] === true) {
+                document.getElementById('notPosTelefono').checked = true;
+                document.getElementById('notNegTelefono').checked = false;
+            } else {
+                document.getElementById('notPosTelefono').checked = false;
+                document.getElementById('notNegTelefono').checked = true;
+            }
         },
         error: function (r) {
             console.log(r);
